requirements.txt
asgiref==3.8.1
Django==4.2.16
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
django-smart-selects==1.6.0
djangorestframework==3.15.2
drf-yasg==1.21.8
inflection==0.5.1
mssql-django==1.5
packaging==24.1
pillow==11.0.0
pyodbc==5.2.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
